# docs/api/openapi.yaml
...
openapi: 3.0.3
...
info:
  title: Smart Tourism Data Platform API
  description: |
    **Enterprise-grade API cho Smart Tourism Data Platform**
    
    API này cung cấp các endpoints để:
    - Quản lý data pipelines (bronze → silver → gold)
    - Query và search POI (Points of Interest) data
    - Giám sát system health và metrics
    - Quản lý data quality và validation
    
    ## Authentication
    
    API sử dụng JWT Bearer token authentication. Thêm header sau vào mỗi request:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Rate Limiting
    
    - 60 requests/minute cho mỗi IP/user (default)
    - 1000 requests/hour cho mỗi authenticated user
    
    ## Data Layers
    
    - **Bronze**: Raw data từ OSM, Google Places, v.v.
    - **Silver**: Cleaned và normalized data
    - **Gold**: Enriched và deduplicated master records
    
  version: 1.0.0
  contact:
    name: Smart Tourism Team
    email: support@smarttravel.vn
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

# ============================================================================
# SERVERS
# ============================================================================
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://staging-api.smarttravel.vn
    description: Staging server
  - url: https://api.smarttravel.vn
    description: Production server

...
# SECURITY
...
security:
  - BearerAuth: []

# ============================================================================
# TAGS
# ============================================================================
...
paths:
...
  /api/v1/pipeline/start:
    post:
      tags:
        - Pipeline Management
      summary: Start pipeline
      description: Khởi động data pipeline với cấu hình được chỉ định
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PipelineStartRequest'
      responses:
        '202':
          description: Pipeline started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStartResponse'
        '400':
          description: Invalid configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Pipeline already running
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pipeline/stop/{pipeline_id}:
    post:
      tags:
        - Pipeline Management
      summary: Stop pipeline
      description: Dừng pipeline đang chạy
      parameters:
        - name: pipeline_id
          in: path
          required: true
          schema:
            type: string
          description: Pipeline ID cần dừng
      responses:
        '200':
          description: Pipeline stopped successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStopResponse'
        '404':
          description: Pipeline not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pipeline/status/{execution_id}:
    get:
      tags:
        - Pipeline Management
      summary: Get pipeline status
      description: Lấy trạng thái và thông tin pipeline execution
      parameters:
        - name: execution_id
          in: path
          required: true
          schema:
            type: string
          description: Execution ID
      responses:
        '200':
          description: Pipeline status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineStatusResponse'
        '404':
          description: Execution not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pipeline/history:
    get:
      tags:
        - Pipeline Management
      summary: Get pipeline history
      description: Lấy lịch sử các pipeline executions
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
          description: Số records trả về
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Offset cho pagination
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, cancelled]
          description: Filter theo status
      responses:
        '200':
          description: Pipeline history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PipelineHistoryResponse'
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token cho authentication

  # ========================================================================
  # SCHEMAS
  # ========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details
      required:
        - error
        - message

    # Health Status
...
    PipelineStartRequest:
      type: object
      properties:
        pipeline_id:
          type: string
        config:
          type: object
        city:
          type: string
        category:
          type: string
        dry_run:
          type: boolean
          default: false

    # Pipeline Start Response
    PipelineStartResponse:
      type: object
      properties:
        execution_id:
          type: string
        pipeline_id:
          type: string
        status:
          type: string
          enum: [pending, running]
        message:
          type: string
        started_at:
          type: string
          format: date-time

    # Pipeline Status Response
    PipelineStatusResponse:
      type: object
      properties:
        execution_id:
          type: string
        pipeline_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed, cancelled]
        progress:
          type: number
          minimum: 0
          maximum: 100
        current_stage:
          type: string
        stages:
          type: array
          items:
            type: object
        metrics:
          type: object
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time

    # Dashboard Response
...